fixed some broken tests due to wrong class due to jsonlite change

+++ tests/testthat/test-dataset_metrics.r
@@ -4,7 +4,7 @@ tt <- dataset_metrics(uuid='3f8a1297-3259-4700-91fc-acc4170b27ce')
 
 test_that("returns the correct class", {
   expect_is(tt, "list")
-  expect_is(tt$key, "numeric")
+  expect_is(tt$key, "integer")
   expect_is(tt$datasetKey, "character")
 })
