+++ tests/testthat/test-name_usage.r
@@ -5,12 +5,12 @@ uu <- name_usage(key=3119195, data='references')
 
 test_that("name_usage returns the correct class", {
   expect_is(tt, "list")
-  expect_is(tt$key, "numeric")
+  expect_is(tt$key, "integer")
   expect_is(tt$kingdom, "character")
   
   expect_is(uu, "list")
   expect_is(uu$results, "list")
-  expect_is(uu$results[[1]]$sourceTaxonKey, "numeric")
+  expect_is(uu$results[[1]]$sourceTaxonKey, "integer")
   expect_is(uu$results[[1]]$citation, "character")
 })
